Combine DiD estimates across cohorts and event times.

Estimate DiD for all possible cohorts and event time pairs (g,e), as well as the average across cohorts for each event time (e).

Usage

DiD(
  inputdata,
  varnames,
  control_group = "all",
  base_event = -1,
  min_event = NULL,
  max_event = NULL,
  Esets = NULL,
  return_ATTs_only = TRUE,
  parallel_cores = 1
)

Arguments

inputdata

A data.table.

varnames

A list of the form varnames = list(id_name, time_name, outcome_name, cohort_name), where all four arguments of the list must be a character that corresponds to a variable name in inputdata.

control_group

There are three possibilities: control_group="never-treated" uses the never-treated control group only; control_group="future-treated" uses those units that will receive treatment in the future as the control group; and control_group="all" uses both the never-treated and the future-treated in the control group. Default is control_group="all".

base_event

This is the base pre-period that is normalized to zero in the DiD estimation. Default is base_event=-1.

min_event

This is the minimum event time (e) to estimate. Default is NULL, in which case, no minimum is imposed.

max_event

This is the maximum event time (e) to estimate. Default is NULL, in which case, no maximum is imposed.

Esets

If a list of sets of event times is provided, it will loop over those sets, computing the average ATT_e across event times e. Default is NULL.

return_ATTs_only

Return only the ATT estimates and sample sizes. Default is TRUE.

parallel_cores

Number of cores to use in parallel processing. If greater than 1, it will try to run library(parallel), so the "parallel" package must be installed. Default is 1.

Value

A list with two components: results_cohort is a data.table with the DiDge estimates (by event e and cohort g), and results_average is a data.table with the DiDe estimates (by event e, average across cohorts g). If the Esets argument is specified, a third component called results_Esets will be included in the list of output.

Examples

# ------ basic usage ------

# simulate some data
simdata = SimDiD(sample_size=1000, ATTcohortdiff = 2)$simdata

# define the variable names as a list()
varnames = list()
varnames$time_name = "year"
varnames$outcome_name = "Y"
varnames$cohort_name = "cohort"
varnames$id_name = "id"

# estimate the ATT for all cohorts at event time 1 only
DiD(simdata, varnames, min_event=1, max_event=1)
#> $results_cohort
#>    Cohort EventTime BaseEvent CalendarTime    ATTge  ATTge_SE Ncontrol Ntreated
#> 1:   2007         1        -1         2008 2.080520 0.1113721      751      249
#> 2:   2010         1        -1         2011 4.074317 0.1066090      501      250
#> 3:   2012         1        -1         2013 6.090972 0.1207522      251      250
#> 
#> $results_average
#>    EventTime BaseEvent     ATTe    ATTe_SE Ncontrol Ntreated
#> 1:         1        -1 4.084608 0.06615348     1503      749
#> 

# change the base period to -3
DiD(simdata, varnames, base_event=-3, min_event=1, max_event=1)
#> $results_cohort
#>    Cohort EventTime BaseEvent CalendarTime    ATTge  ATTge_SE Ncontrol Ntreated
#> 1:   2007         1        -3         2008 2.202747 0.1011322      751      249
#> 2:   2010         1        -3         2011 3.873898 0.1090590      501      250
#> 3:   2012         1        -3         2013 6.159048 0.1361621      251      250
#> 
#> $results_average
#>    EventTime BaseEvent     ATTe    ATTe_SE Ncontrol Ntreated
#> 1:         1        -3 4.081069 0.06731803     1503      749
#> 

# check the pre-periods -4 through -2
DiD(simdata, varnames, control_group = "never-treated", min_event=-4, max_event=-2)
#> $results_cohort
#>    Cohort EventTime BaseEvent CalendarTime        ATTge  ATTge_SE Ncontrol
#> 1:   2007        -4        -1         2003 -0.006368097 0.1265188      251
#> 2:   2007        -3        -1         2004 -0.001150727 0.1280774      251
#> 3:   2007        -2        -1         2005  0.185373922 0.1248511      251
#> 4:   2010        -4        -1         2006  0.092126680 0.1310240      251
#> 5:   2010        -3        -1         2007  0.158665767 0.1320550      251
#> 6:   2010        -2        -1         2008  0.287904081 0.1258577      251
#> 7:   2012        -4        -1         2008  0.133280149 0.1207528      251
#> 8:   2012        -3        -1         2009 -0.068076369 0.1330084      251
#> 9:   2012        -2        -1         2010 -0.128911662 0.1210347      251
#>    Ntreated
#> 1:      249
#> 2:      249
#> 3:      249
#> 4:      250
#> 5:      250
#> 6:      250
#> 7:      250
#> 8:      250
#> 9:      250
#> 
#> $results_average
#>    EventTime BaseEvent       ATTe    ATTe_SE Ncontrol Ntreated
#> 1:        -4        -1 0.07311889 0.06736954      753      749
#> 2:        -3        -1 0.02985423 0.06616998      753      749
#> 3:        -2        -1 0.11469454 0.07116391      753      749
#> 

# use only the never-treated control group, estimate events -4 through 1
DiD(simdata, varnames, control_group = "never-treated", min_event=-4, max_event=1)
#> $results_cohort
#>     Cohort EventTime BaseEvent CalendarTime        ATTge  ATTge_SE Ncontrol
#>  1:   2007        -4        -1         2003 -0.006368097 0.1265188      251
#>  2:   2007        -3        -1         2004 -0.001150727 0.1280774      251
#>  3:   2007        -2        -1         2005  0.185373922 0.1248511      251
#>  4:   2007        -1        -1         2006  0.000000000 0.0000000      251
#>  5:   2007         0        -1         2007  0.952907181 0.1291093      251
#>  6:   2007         1        -1         2008  2.236676273 0.1303068      251
#>  7:   2010        -4        -1         2006  0.092126680 0.1310240      251
#>  8:   2010        -3        -1         2007  0.158665767 0.1320550      251
#>  9:   2010        -2        -1         2008  0.287904081 0.1258577      251
#> 10:   2010        -1        -1         2009  0.000000000 0.0000000      251
#> 11:   2010         0        -1         2010  2.983597405 0.1286926      251
#> 12:   2010         1        -1         2011  4.108286869 0.1264383      251
#> 13:   2012        -4        -1         2008  0.133280149 0.1207528      251
#> 14:   2012        -3        -1         2009 -0.068076369 0.1330084      251
#> 15:   2012        -2        -1         2010 -0.128911662 0.1210347      251
#> 16:   2012        -1        -1         2011  0.000000000 0.0000000      251
#> 17:   2012         0        -1         2012  4.881242047 0.1282893      251
#> 18:   2012         1        -1         2013  6.090971782 0.1207522      251
#>     Ntreated
#>  1:      249
#>  2:      249
#>  3:      249
#>  4:      249
#>  5:      249
#>  6:      249
#>  7:      250
#>  8:      250
#>  9:      250
#> 10:      250
#> 11:      250
#> 12:      250
#> 13:      250
#> 14:      250
#> 15:      250
#> 16:      250
#> 17:      250
#> 18:      250
#> 
#> $results_average
#>    EventTime BaseEvent       ATTe    ATTe_SE Ncontrol Ntreated
#> 1:        -4        -1 0.07311889 0.06736954      753      749
#> 2:        -3        -1 0.02985423 0.06616998      753      749
#> 3:        -2        -1 0.11469454 0.07116391      753      749
#> 4:        -1        -1 0.00000000 0.00000000      753      749
#> 5:         0        -1 2.94190087 0.07480281      753      749
#> 6:         1        -1 4.14785989 0.06824073      753      749
#> 

# estimate average ATTe across sets of events
DiD(simdata, varnames, min_event=-4, max_event=6, Esets=list(c(-4,-3,-2),c(1,2,3)))
#> $results_cohort
#>     Cohort EventTime BaseEvent CalendarTime       ATTge  ATTge_SE Ncontrol
#>  1:   2007        -4        -1         2003 -0.04000461 0.1063722      751
#>  2:   2007        -3        -1         2004 -0.12222638 0.1067716      751
#>  3:   2007        -2        -1         2005  0.09127603 0.1032551      751
#>  4:   2007        -1        -1         2006  0.00000000 0.0000000      751
#>  5:   2007         0        -1         2007  0.93465627 0.1059910      751
#>  6:   2007         1        -1         2008  2.08052015 0.1113721      751
#>  7:   2007         2        -1         2009  2.89977611 0.1046816      751
#>  8:   2007         3        -1         2010  3.94963207 0.1143283      501
#>  9:   2007         4        -1         2011  5.06524420 0.1152643      501
#> 10:   2007         5        -1         2012  6.04106888 0.1238934      251
#> 11:   2007         6        -1         2013  7.00871764 0.1287452      251
#> 12:   2010        -4        -1         2006  0.12803440 0.1103934      501
#> 13:   2010        -3        -1         2007  0.20041848 0.1091284      501
#> 14:   2010        -2        -1         2008  0.18742678 0.1064162      501
#> 15:   2010        -1        -1         2009  0.00000000 0.0000000      501
#> 16:   2010         0        -1         2010  3.01395434 0.1089320      501
#> 17:   2010         1        -1         2011  4.07431662 0.1066090      501
#> 18:   2010         2        -1         2012  5.02530306 0.1316073      251
#> 19:   2010         3        -1         2013  6.11411205 0.1290900      251
#> 20:   2012        -4        -1         2008  0.13328015 0.1207528      251
#> 21:   2012        -3        -1         2009 -0.06807637 0.1330084      251
#> 22:   2012        -2        -1         2010 -0.12891166 0.1210347      251
#> 23:   2012        -1        -1         2011  0.00000000 0.0000000      251
#> 24:   2012         0        -1         2012  4.88124205 0.1282893      251
#> 25:   2012         1        -1         2013  6.09097178 0.1207522      251
#>     Cohort EventTime BaseEvent CalendarTime       ATTge  ATTge_SE Ncontrol
#>     Ntreated
#>  1:      249
#>  2:      249
#>  3:      249
#>  4:      249
#>  5:      249
#>  6:      249
#>  7:      249
#>  8:      249
#>  9:      249
#> 10:      249
#> 11:      249
#> 12:      250
#> 13:      250
#> 14:      250
#> 15:      250
#> 16:      250
#> 17:      250
#> 18:      250
#> 19:      250
#> 20:      250
#> 21:      250
#> 22:      250
#> 23:      250
#> 24:      250
#> 25:      250
#>     Ntreated
#> 
#> $results_average
#>     EventTime BaseEvent        ATTe    ATTe_SE Ncontrol Ntreated
#>  1:        -4        -1 0.073921885 0.06558170     1503      749
#>  2:        -3        -1 0.003539596 0.06519249     1503      749
#>  3:        -2        -1 0.049875181 0.06320729     1503      749
#>  4:        -1        -1 0.000000000 0.00000000     1503      749
#>  5:         0        -1 2.945965965 0.06603599     1503      749
#>  6:         1        -1 4.084608302 0.06615348     1503      749
#>  7:         2        -1 3.964669375 0.08260010     1002      499
#>  8:         3        -1 5.034040875 0.08637273      752      499
#>  9:         4        -1 5.065244197 0.11526433      501      249
#> 10:         5        -1 6.041068883 0.12389338      251      249
#> 11:         6        -1 7.008717639 0.12874515      251      249
#> 
#> $results_Esets
#>        Eset   ATT_Eset ATT_Eset_SE
#> 1: -4,-3,-2 0.04244555  0.05245430
#> 2:    1,2,3 4.18897766  0.06019071
#>