DiD data simulator with staggered treatment.

Simulate data from the model Y_it = alpha_i + mu_t + ATT*(t >= G_i) + epsilon_it, where i is individual, t is year, and G_i is the cohort. The ATT formula is ATTat0 + EventTime*ATTgrowth + \*cohort_counter\*ATTcohortdiff, where cohort_counter is the order of treated cohort (first, second, etc.).

Usage

SimDiD(
  seed = 1,
  sample_size = 100,
  cohorts = c(2007, 2010, 2012),
  ATTat0 = 1,
  ATTgrowth = 1,
  ATTcohortdiff = 0.5,
  anticipation = 0,
  minyear = 2003,
  maxyear = 2013,
  idvar = 1,
  yearvar = 1,
  shockvar = 1,
  indivAR1 = FALSE,
  time_covars = FALSE,
  clusters = FALSE,
  markets = FALSE,
  randomNA = FALSE,
  missingCohorts = NULL
)

Arguments

seed

Set the random seed. Default is seed=1.

sample_size

Number of individuals. Default is sample_size=100.

cohorts

Vector of years at which treatment onset occurs. Default is cohorts=c(2007,2010,2012).

ATTat0

Treatment effect at event time 0. Default is 1.

ATTgrowth

Increment in the ATT for each event time after 0. Default is 1.

ATTcohortdiff

Incrememnt in the ATT for each cohort. Default is 0.5.

anticipation

Number of years prior to cohort to allow 50% treatment effects. Default is anticipation=0.

minyear

Minimum calendar year to include in the data. Default is minyear=2003.

maxyear

Maximum calendar year to include in the data. Default is maxyear=2013.

idvar

Variance of individual fixed effects (alpha_i). Default is idvar=1.

yearvar

Variance of year effects (mu_i). Default is yearvar=1.

shockvar

Variance of idiosyncratic shocks (epsilon_it). Default is shockvar=1.

indivAR1

Each individual's shocks follow an AR(1) process. Default is FALSE.

time_covars

Add 2 time-varying covariates, called "X1" and "X2". Default is FALSE.

clusters

Add 10 randomly assigned clusters, with cluster-specific AR(1) shocks. Default is FALSE.

markets

Add 10 randomly assigned markets, with market-specific shocks that are systematically greater for markets that are treated earlier. Default is FALSE.

randomNA

If TRUE, randomly assign the outcome variable with missing values (NA) in some cases. Default is FALSE.

missingCohorts

If set to a particular cohort (or vector of cohorts), all of the outcomes for that cohort at event time -1 will be set to missing. Default is NULL.

Value

A list with two data.tables. The first data.table is simulated data with variables (id, year, cohort, Y), where Y is the outcome variable. The second data.table contains the true ATT values, both at the (event,cohort) level and by event averaging across cohorts.

Examples

# simulate data with default options
SimDiD()
#> $simdata
#>        id year cohort         Y
#>    1:   1 2003   2012  8.058406
#>    2:   1 2004   2012 12.348703
#>    3:   1 2005   2012  7.549438
#>    4:   1 2006   2012  9.731058
#>    5:   1 2007   2012 11.269052
#>   ---                          
#> 1096: 100 2009   2010  7.071765
#> 1097: 100 2010   2010 12.349172
#> 1098: 100 2011   2010 13.647523
#> 1099: 100 2012   2010 13.490701
#> 1100: 100 2013   2010 12.650051
#> 
#> $true_ATT
#>      cohort event    ATTge
#>  1:    2007     0 1.000000
#>  2:    2007     1 2.000000
#>  3:    2007     2 3.000000
#>  4:    2007     3 4.000000
#>  5:    2007     4 5.000000
#>  6:    2007     5 6.000000
#>  7:    2007     6 7.000000
#>  8:    2010     0 1.500000
#>  9:    2010     1 2.500000
#> 10:    2010     2 3.500000
#> 11:    2010     3 4.500000
#> 12:    2012     0 2.000000
#> 13:    2012     1 3.000000
#> 14: Average     0 1.506757
#> 15: Average     1 2.506757
#> 16: Average     2 3.255102
#> 17: Average     3 4.255102
#> 18: Average     4 5.000000
#> 19: Average     5 6.000000
#> 20: Average     6 7.000000
#> 

# change the random seed
SimDiD(seed=123)
#> $simdata
#>        id year cohort         Y
#>    1:   1 2003    Inf 10.975366
#>    2:   1 2004    Inf  9.928304
#>    3:   1 2005    Inf 13.326225
#>    4:   1 2006    Inf 10.238570
#>    5:   1 2007    Inf 11.675520
#>   ---                          
#> 1096: 100 2009   2012  9.828756
#> 1097: 100 2010   2012 11.920037
#> 1098: 100 2011   2012 10.924002
#> 1099: 100 2012   2012 13.151364
#> 1100: 100 2013   2012 13.329865
#> 
#> $true_ATT
#>      cohort event    ATTge
#>  1:    2007     0 1.000000
#>  2:    2007     1 2.000000
#>  3:    2007     2 3.000000
#>  4:    2007     3 4.000000
#>  5:    2007     4 5.000000
#>  6:    2007     5 6.000000
#>  7:    2007     6 7.000000
#>  8:    2010     0 1.500000
#>  9:    2010     1 2.500000
#> 10:    2010     2 3.500000
#> 11:    2010     3 4.500000
#> 12:    2012     0 2.000000
#> 13:    2012     1 3.000000
#> 14: Average     0 1.506757
#> 15: Average     1 2.506757
#> 16: Average     2 3.255102
#> 17: Average     3 4.255102
#> 18: Average     4 5.000000
#> 19: Average     5 6.000000
#> 20: Average     6 7.000000
#> 

# increase sample size
SimDiD(sample_size=200)
#> $simdata
#>        id year cohort         Y
#>    1:   1 2003    Inf  9.916155
#>    2:   1 2004    Inf 12.095935
#>    3:   1 2005    Inf 10.183453
#>    4:   1 2006    Inf 13.995163
#>    5:   1 2007    Inf 11.806494
#>   ---                          
#> 2196: 200 2009   2007 15.742248
#> 2197: 200 2010   2007 14.753802
#> 2198: 200 2011   2007 15.422272
#> 2199: 200 2012   2007 15.432962
#> 2200: 200 2013   2007 17.223699
#> 
#> $true_ATT
#>      cohort event    ATTge
#>  1:    2007     0 1.000000
#>  2:    2007     1 2.000000
#>  3:    2007     2 3.000000
#>  4:    2007     3 4.000000
#>  5:    2007     4 5.000000
#>  6:    2007     5 6.000000
#>  7:    2007     6 7.000000
#>  8:    2010     0 1.500000
#>  9:    2010     1 2.500000
#> 10:    2010     2 3.500000
#> 11:    2010     3 4.500000
#> 12:    2012     0 2.000000
#> 13:    2012     1 3.000000
#> 14: Average     0 1.503356
#> 15: Average     1 2.503356
#> 16: Average     2 3.252525
#> 17: Average     3 4.252525
#> 18: Average     4 5.000000
#> 19: Average     5 6.000000
#> 20: Average     6 7.000000
#> 

# increase the initial ATT for the first treated cohort
SimDiD(ATTat0=2)
#> $simdata
#>        id year cohort         Y
#>    1:   1 2003   2012  8.058406
#>    2:   1 2004   2012 12.348703
#>    3:   1 2005   2012  7.549438
#>    4:   1 2006   2012  9.731058
#>    5:   1 2007   2012 11.269052
#>   ---                          
#> 1096: 100 2009   2010  7.071765
#> 1097: 100 2010   2010 13.349172
#> 1098: 100 2011   2010 14.647523
#> 1099: 100 2012   2010 14.490701
#> 1100: 100 2013   2010 13.650051
#> 
#> $true_ATT
#>      cohort event    ATTge
#>  1:    2007     0 2.000000
#>  2:    2007     1 3.000000
#>  3:    2007     2 4.000000
#>  4:    2007     3 5.000000
#>  5:    2007     4 6.000000
#>  6:    2007     5 7.000000
#>  7:    2007     6 8.000000
#>  8:    2010     0 2.500000
#>  9:    2010     1 3.500000
#> 10:    2010     2 4.500000
#> 11:    2010     3 5.500000
#> 12:    2012     0 3.000000
#> 13:    2012     1 4.000000
#> 14: Average     0 2.506757
#> 15: Average     1 3.506757
#> 16: Average     2 4.255102
#> 17: Average     3 5.255102
#> 18: Average     4 6.000000
#> 19: Average     5 7.000000
#> 20: Average     6 8.000000
#> 

# change cross-time and cross-cohort ATT variation
# note: the formula is ATTat0 + EventTime*ATTgrowth
SimDiD(ATTgrowth=2,ATTcohortdiff=3)
#> $simdata
#>        id year cohort         Y
#>    1:   1 2003   2012  8.058406
#>    2:   1 2004   2012 12.348703
#>    3:   1 2005   2012  7.549438
#>    4:   1 2006   2012  9.731058
#>    5:   1 2007   2012 11.269052
#>   ---                          
#> 1096: 100 2009   2010  7.071765
#> 1097: 100 2010   2010 14.849172
#> 1098: 100 2011   2010 17.147523
#> 1099: 100 2012   2010 17.990701
#> 1100: 100 2013   2010 18.150051
#> 
#> $true_ATT
#>      cohort event     ATTge
#>  1:    2007     0  1.000000
#>  2:    2007     1  3.000000
#>  3:    2007     2  5.000000
#>  4:    2007     3  7.000000
#>  5:    2007     4  9.000000
#>  6:    2007     5 11.000000
#>  7:    2007     6 13.000000
#>  8:    2010     0  4.000000
#>  9:    2010     1  6.000000
#> 10:    2010     2  8.000000
#> 11:    2010     3 10.000000
#> 12:    2012     0  7.000000
#> 13:    2012     1  9.000000
#> 14: Average     0  4.040541
#> 15: Average     1  6.040541
#> 16: Average     2  6.530612
#> 17: Average     3  8.530612
#> 18: Average     4  9.000000
#> 19: Average     5 11.000000
#> 20: Average     6 13.000000
#> 

# change the range of years considered and treatment cohorts
SimDiD(minyear=1994, maxyear=1999, cohorts=c(1995,1996))
#> $simdata
#>       id year cohort         Y
#>   1:   1 1994   1996  8.324911
#>   2:   1 1995   1996 11.814148
#>   3:   1 1996   1996 10.233306
#>   4:   1 1997   1996 13.740197
#>   5:   1 1998   1996 12.364984
#>  ---                          
#> 596: 100 1995   1996 12.435341
#> 597: 100 1996   1996 10.289332
#> 598: 100 1997   1996 11.745367
#> 599: 100 1998   1996 12.976739
#> 600: 100 1999   1996 12.088275
#> 
#> $true_ATT
#>      cohort event ATTge
#>  1:    1995     0  1.00
#>  2:    1995     1  2.00
#>  3:    1995     2  3.00
#>  4:    1995     3  4.00
#>  5:    1995     4  5.00
#>  6:    1996     0  1.50
#>  7:    1996     1  2.50
#>  8:    1996     2  3.50
#>  9:    1996     3  4.50
#> 10: Average     0  1.25
#> 11: Average     1  2.25
#> 12: Average     2  3.25
#> 13: Average     3  4.25
#> 14: Average     4  5.00
#> 

# add two periods of anticipation prior to treatment. The effect will be 0.5*ATTat0.
SimDiD(anticipation=2)
#> $simdata
#>        id year cohort         Y
#>    1:   1 2003   2012  8.058406
#>    2:   1 2004   2012 12.348703
#>    3:   1 2005   2012  7.549438
#>    4:   1 2006   2012  9.731058
#>    5:   1 2007   2012 11.269052
#>   ---                          
#> 1096: 100 2009   2010  7.571765
#> 1097: 100 2010   2010 12.349172
#> 1098: 100 2011   2010 13.647523
#> 1099: 100 2012   2010 13.490701
#> 1100: 100 2013   2010 12.650051
#> 
#> $true_ATT
#>      cohort event    ATTge
#>  1:    2007    -2 0.500000
#>  2:    2007    -1 0.500000
#>  3:    2007     0 1.000000
#>  4:    2007     1 2.000000
#>  5:    2007     2 3.000000
#>  6:    2007     3 4.000000
#>  7:    2007     4 5.000000
#>  8:    2007     5 6.000000
#>  9:    2007     6 7.000000
#> 10:    2010    -2 0.500000
#> 11:    2010    -1 0.500000
#> 12:    2010     0 1.500000
#> 13:    2010     1 2.500000
#> 14:    2010     2 3.500000
#> 15:    2010     3 4.500000
#> 16:    2012    -2 0.500000
#> 17:    2012    -1 0.500000
#> 18:    2012     0 2.000000
#> 19:    2012     1 3.000000
#> 20: Average    -2 0.500000
#> 21: Average    -1 0.500000
#> 22: Average     0 1.506757
#> 23: Average     1 2.506757
#> 24: Average     2 3.255102
#> 25: Average     3 4.255102
#> 26: Average     4 5.000000
#> 27: Average     5 6.000000
#> 28: Average     6 7.000000
#>      cohort event    ATTge
#> 

# add covariates "X1" and "X2", which vary over time differentially by cohort.
SimDiD(time_covars=TRUE)
#> $simdata
#>        id year cohort         Y          X1          X2
#>    1:   1 2003   2012 14.256933 -3.69550292  2.50302353
#>    2:   1 2004   2012 17.874436 -3.80398827  1.72174457
#>    3:   1 2005   2012 12.261336 -3.33934345  1.37255519
#>    4:   1 2006   2012 12.064172 -0.73485586  1.59825808
#>    5:   1 2007   2012 12.691232 -0.06022631  1.36195377
#>   ---                                                  
#> 1096: 100 2009   2010  5.871112  1.18163056 -0.01902207
#> 1097: 100 2010   2010 11.615832  1.93069848  1.19735833
#> 1098: 100 2011   2010 11.933958  2.70553461  0.99196922
#> 1099: 100 2012   2010  6.879272  5.41403034 -1.19739883
#> 1100: 100 2013   2010  9.111691  6.09999773  2.56163746
#> 
#> $true_ATT
#>      cohort event    ATTge
#>  1:    2007     0 1.000000
#>  2:    2007     1 2.000000
#>  3:    2007     2 3.000000
#>  4:    2007     3 4.000000
#>  5:    2007     4 5.000000
#>  6:    2007     5 6.000000
#>  7:    2007     6 7.000000
#>  8:    2010     0 1.500000
#>  9:    2010     1 2.500000
#> 10:    2010     2 3.500000
#> 11:    2010     3 4.500000
#> 12:    2012     0 2.000000
#> 13:    2012     1 3.000000
#> 14: Average     0 1.506757
#> 15: Average     1 2.506757
#> 16: Average     2 3.255102
#> 17: Average     3 4.255102
#> 18: Average     4 5.000000
#> 19: Average     5 6.000000
#> 20: Average     6 7.000000
#> 

# add clusters called "cluster", which are time-invariant and subject to common shocks.
# note: shocks are randomly assigned to clusters, so parallel trends is not violated.
SimDiD(clusters=TRUE)
#> $simdata
#>        id year cohort         Y cluster
#>    1:   1 2003   2012  9.009133       2
#>    2:   1 2004   2012 14.009706       2
#>    3:   1 2005   2012 12.019513       2
#>    4:   1 2006   2012 11.912093       2
#>    5:   1 2007   2012 11.691422       2
#>   ---                                  
#> 1096: 100 2009   2010 10.062813       8
#> 1097: 100 2010   2010 13.625028       8
#> 1098: 100 2011   2010 13.245403       8
#> 1099: 100 2012   2010 13.482591       8
#> 1100: 100 2013   2010 12.090760       8
#> 
#> $true_ATT
#>      cohort event    ATTge
#>  1:    2007     0 1.000000
#>  2:    2007     1 2.000000
#>  3:    2007     2 3.000000
#>  4:    2007     3 4.000000
#>  5:    2007     4 5.000000
#>  6:    2007     5 6.000000
#>  7:    2007     6 7.000000
#>  8:    2010     0 1.500000
#>  9:    2010     1 2.500000
#> 10:    2010     2 3.500000
#> 11:    2010     3 4.500000
#> 12:    2012     0 2.000000
#> 13:    2012     1 3.000000
#> 14: Average     0 1.506757
#> 15: Average     1 2.506757
#> 16: Average     2 3.255102
#> 17: Average     3 4.255102
#> 18: Average     4 5.000000
#> 19: Average     5 6.000000
#> 20: Average     6 7.000000
#> 

# add clusters called "market", which are time-invariant and subject to common shocks.
# note: earlier-treated markets tend to get better shocks, which violates parallel trends.
SimDiD(markets=TRUE)
#> $simdata
#>        id year cohort         Y market
#>    1:   1 2003   2012  7.858406      2
#>    2:   1 2004   2012 12.148703      2
#>    3:   1 2005   2012  7.349438      2
#>    4:   1 2006   2012  9.531058      2
#>    5:   1 2007   2012 11.069052      2
#>   ---                                 
#> 1096: 100 2009   2010  5.471765      8
#> 1097: 100 2010   2010 10.749172      8
#> 1098: 100 2011   2010 12.047523      8
#> 1099: 100 2012   2010 11.890701      8
#> 1100: 100 2013   2010 11.050051      8
#> 
#> $true_ATT
#>      cohort event    ATTge
#>  1:    2007     0 1.000000
#>  2:    2007     1 2.000000
#>  3:    2007     2 3.000000
#>  4:    2007     3 4.000000
#>  5:    2007     4 5.000000
#>  6:    2007     5 6.000000
#>  7:    2007     6 7.000000
#>  8:    2010     0 1.500000
#>  9:    2010     1 2.500000
#> 10:    2010     2 3.500000
#> 11:    2010     3 4.500000
#> 12:    2012     0 2.000000
#> 13:    2012     1 3.000000
#> 14: Average     0 1.506757
#> 15: Average     1 2.506757
#> 16: Average     2 3.255102
#> 17: Average     3 4.255102
#> 18: Average     4 5.000000
#> 19: Average     5 6.000000
#> 20: Average     6 7.000000
#> 

# randomly set 10% of outcomes to missing.
SimDiD(randomNA=TRUE)
#> $simdata
#>        id year cohort         Y
#>    1:   1 2003   2012  8.058406
#>    2:   1 2004   2012 12.348703
#>    3:   1 2005   2012  7.549438
#>    4:   1 2006   2012  9.731058
#>    5:   1 2007   2012 11.269052
#>   ---                          
#> 1096: 100 2009   2010  7.071765
#> 1097: 100 2010   2010 12.349172
#> 1098: 100 2011   2010 13.647523
#> 1099: 100 2012   2010 13.490701
#> 1100: 100 2013   2010 12.650051
#> 
#> $true_ATT
#>      cohort event    ATTge
#>  1:    2007     0 1.000000
#>  2:    2007     1 2.000000
#>  3:    2007     2 3.000000
#>  4:    2007     3 4.000000
#>  5:    2007     4 5.000000
#>  6:    2007     5 6.000000
#>  7:    2007     6 7.000000
#>  8:    2010     0 1.500000
#>  9:    2010     1 2.500000
#> 10:    2010     2 3.500000
#> 11:    2010     3 4.500000
#> 12:    2012     0 2.000000
#> 13:    2012     1 3.000000
#> 14: Average     0 1.506757
#> 15: Average     1 2.506757
#> 16: Average     2 3.255102
#> 17: Average     3 4.255102
#> 18: Average     4 5.000000
#> 19: Average     5 6.000000
#> 20: Average     6 7.000000
#> 

# set cohort 2010 outcome to missing at event time -1.
SimDiD(missingCohorts=2010)$simdata[cohort==2010 & year==2009]
#>      id year cohort  Y
#>  1:  18 2009   2010 NA
#>  2:  22 2009   2010 NA
#>  3:  23 2009   2010 NA
#>  4:  24 2009   2010 NA
#>  5:  29 2009   2010 NA
#>  6:  30 2009   2010 NA
#>  7:  38 2009   2010 NA
#>  8:  40 2009   2010 NA
#>  9:  46 2009   2010 NA
#> 10:  48 2009   2010 NA
#> 11:  50 2009   2010 NA
#> 12:  60 2009   2010 NA
#> 13:  61 2009   2010 NA
#> 14:  68 2009   2010 NA
#> 15:  70 2009   2010 NA
#> 16:  71 2009   2010 NA
#> 17:  72 2009   2010 NA
#> 18:  77 2009   2010 NA
#> 19:  83 2009   2010 NA
#> 20:  84 2009   2010 NA
#> 21:  90 2009   2010 NA
#> 22:  92 2009   2010 NA
#> 23:  96 2009   2010 NA
#> 24:  98 2009   2010 NA
#> 25: 100 2009   2010 NA
#>      id year cohort  Y